Reviewed all of the links to evidence.eu.com now that the erika site exists.
//////// Here below: what e things I was not completely sure about //// 

///1

The following occurrence was present in the biblio.bib in manuals/common/

"
@Misc{Evidence-multicpu-tutorial,
  author =       {Evidence Srl},
  title =        {ERIKA Enterprise Reference Manual, Chapter 2},
  howpublished = {Evidence literature page, \url{http://www.evidence.eu.com/nios2/literature.asp}},
  year =         {2005}
}
"

the url is now:
http://erika.tuxfamily.org/evidence/documentation.html,
which is the only actually published collection of links to manuals

Other automatically produced documentation html are produced.
Aside from the one used in this published page
(download/manuals/include/dynamic-man.php) also 
dynamic-man-nios2.php is produced (and not published).

!! BUT !!
was not that the original link (one that ends to be in every manual)
other than being outdated, just offtopic?

It is the only link in bibliography regarding Erika...
what does NIOS2 got to do with it?





///2
Of all the changes the only one with a little doubt is this.
For sure here the erika site should be the main one to be cited... but in this
case an additional link to the flex store could be not completely offtopic.

I'm leaving this note here also as a sign on how the other changes were obvious.



Here's the full "scilab_codegen/trunk/ee/scicos_examples/Demoboard_USB/Readme.txt" as it was:

"
This demo shows how to estabilish a communication
among a PC and the FLEX FULL board through a USB connection.

* The file "demoboard_example_flexside.cos" contains the schematic for the generation
  (with SCICOS) of the FLEX FULL board program. With this application the FLEX FULL
  board is able to:
  - acquire data from the acceleromter and from the light sensor;
  - send the measures to a PC using the USB interface;
  - receive and display the USB data on the LCD screen;
  - perform elaborations for a pretty LEDs blinking;

* The file "demoboard_example_pcside.cos" contains the schematic for the PC-side. 
  The almost-real-time simulation runs in SCICOS environment. The application
  shows the received USB data on the multiple scope and sends other data to the
  FLEX FULL board. The simulation requires the simultaneous execution of a program
  named "app_flex_scicos.exe". To compile the executable use
  the makefile and the source files located in the following folder:
  - .\contrib\USB_UDP_FLEX_SCICOS_App

                                                              13,1          Top
For other info go to the site: http://www.evidence.eu.com/
"


git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@400 88d5ba13-baa4-4915-ae7a-41304cef7dd7