PPC e200zX: fixed compilation errors
Removed duplicate source file in makefiles
Updated ee_debug.h, which was not in sync with some recent changes
Workaround for names clashing in libee.a (see below)

There are two files generating an object named ee_context.o in the Erika tree.  If libee.a exists and should be updated, it is possible that the wrong ee_context.o inside the library is replaced.

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1188 88d5ba13-baa4-4915-ae7a-41304cef7dd7