PPC e200zX: fixed problems with VLE
A bad interaction between Erika Lauterbach scripts, assembly directives,
and VLE setting caused Lauterbach debugger to choose the wrong instruction
set and show FLOWERRORS.  Also, flashing from the Lauterbach failed if VLE
were used, because of a bad MMU setting.  Now the VLE flag is generated
correctly in the ELF debug symbols, and the scripts behave as expected.

Warning: old versions of the Lauterbach T32 sowftware may raise errors when
analyzing a trace.

Also, added a workaround to support the `bctr' instruction in the Diab
assembler when VLE is used.

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1202 88d5ba13-baa4-4915-ae7a-41304cef7dd7