AS, PPC e200zX: Rewritten system call wrappers
System call wrappers are real functions, so that they can be defined in a
type-safe way.  The Diab implementation (currently, the only one) consists
in in-line assembly, which avoid the additional costs of a function call.

The wrapper definitions have been generated through two shell scripts.
They can be adapted and reused for other architectures or compilers.

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1389 88d5ba13-baa4-4915-ae7a-41304cef7dd7