Erika e200zx: added comment for the eval_misra makefile
This is just a comment to warn about the fact that the awk script used inside
the makefile of the eval_misra project may cause an error in the output file
(errors.report). This is due to the fact that in some scenarios the Erika tree
pathname may be too long. Another anomaly concerns the environmental variable
EEBASE used inside the makefile. If it is not defined in the context of the
makefile the "/" character is removed from pathname.

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1855 88d5ba13-baa4-4915-ae7a-41304cef7dd7