Erika e200zx, examples: all the examples are ok
All the examples are ok except for "porting_examples".

Porting examples is a collection of cases useful to do a porting. There
are some of these examples that work with the RT-Druid support,
some other that work fine with the fake_druid support and some that work
with both. Since they are not all supported by RT-Druid, they cannot be
considered for automatic tests.

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1864 88d5ba13-baa4-4915-ae7a-41304cef7dd7