Kernel OO: TaskBody and ISR2Body ErrorHook changed
The Autosar OS367 requires that the system service that do not return
a StatusType shall not generate the Error Hook.

Functions EE_IRQ_end_post_stub() and EE_thread_not_terminated()
are not service routines and currently they raise a ErrorHook, but
they do not return a StatusType. Since they are not service routines
a returned StatusType is not required to accomplish with OS367.

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1905 88d5ba13-baa4-4915-ae7a-41304cef7dd7