Erika e200zx, Board: EE_button_get_B0() fixed.
This patch fixes the issue related with the return value of EE_button_get_B0()
function. Currently this function returns 1 if the button of the evaluation
board is pressed, 0 otherwise.

In order to be coherent with the comments present inside the if-else statement
the "!" operator has been added inside the if-statement, hence if the button is
pressed the else-statement is executed, if the button is not pressed the
if-statement is executed.

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1979 88d5ba13-baa4-4915-ae7a-41304cef7dd7