kernel OO: AR requirement OS093 about call service with interrupts disabled.
If interrupts are disabled/suspended by a Task/OsIsr and the Task/OsIsr calls
any OS service (excluding the interrupt services) then the Operating System
shall ignore the service AND shall return E_OS_DISABLEDINT if the service
returns a StatusType value.

Currently the solution rely only in internal kernel status
(EE_oo_IRQ_disable_count != 0U) and do not actually check the hardware because
there is not hal support for that yet.

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@1985 88d5ba13-baa4-4915-ae7a-41304cef7dd7