e200zx: Modified makefile to let the user obtain Erika as lib.
Added support for EE_OPT="__NO_APP__"; (as cortex and pother do) to stop
compilation at the creation of libee.a (no APP_SRC linked with it).

Added support EE_OPT = "EE_CONFIGURATED_LIB"; if the user want the lib already
configured (that means eecfg.c added as source to the lib).

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@2026 88d5ba13-baa4-4915-ae7a-41304cef7dd7