Erika e200zx, MPC5668 multicore: MPC5668G modified due to the MPC5643L
This patch has impact on MPC5668G due to the introduction of MPC5643L
(leopard). In particular on the external priority assignement.

This patch also modified the system start symbol: __start becomes ee_start
due to the fact that the symbol __start conflicts witrh the same symbol
defined in Codewarrior run-time libaries.

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@2039 88d5ba13-baa4-4915-ae7a-41304cef7dd7