Erika Leopard, Board support: added enable/disable button interrupt
This patch adds the following functions to allow enable/disable
button interrupts:

1) EE_buttons_disable_interrupts
2) EE_buttons_enable_interrupts

Please notice that functions above has a unusefull input parameter. It is
provided with the only intent to guarantee backward comnpatibility with
MAMBA (MPC5674F) evaluation board. The parameter will be removed in the
future release.

NOTE: this release of Erika has been tested with the following ARCHs
of the modistarc testcases:
1) ARCH=e200zx_codewarrior_10_0_2_vle
2) ARCH=e200zx_codewarrior_10_0_2_fle

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@2044 88d5ba13-baa4-4915-ae7a-41304cef7dd7