Erika e200zx multicore: added new type EE_TYPESPINSTATUS
This patch adds new type to allow RT-Druid providing the same generated
code for different architectures (using and not using this type).

Notice that this type is not still used in e200zx, it has been provided
just to guarantee RT-Druid a correct behaviour for different multicore
erchitectures. Recent version of RT-druid generates a data structure
having this type, if a particular architecture supports such structure,
it will be used by the kernel, otherwise it should be stripped by the
compiler since it is not used (but the build process is guaranteed).

git-svn-id: svn://svn.tuxfamily.org/svnroot/erika/erikae/repos/ee/trunk/ee@2806 88d5ba13-baa4-4915-ae7a-41304cef7dd7